Refactor Trainer Class and Introduce ejit for Enhanced Performance
- Renamed `create_collect_function` to `create_grain_collect_function` for clarity.
- Added `create_tfds_collect_function` to handle TensorFlow datasets.
- Updated `create_collect_function` to conditionally use grain or TensorFlow datasets based on `use_grain` argument.
- Replaced `jax.jit` with `ejit` for JIT compilation, improving caching and performance.
- Updated `BaseTrainerProtocol` to include abstract methods for grain and TensorFlow dataset collection functions.
- Modified `TrainingArguments` to include `use_grain` flag for selecting data collection method.
- Refactored `compiling_utils.py` to implement `ejit`, a caching decorator for JAX functions.
- Updated tests to utilize `ejit` instead of `jax.jit` for consistency and performance.
- Added GitHub Actions workflow for linting the codebase using Ruff.

+++ tests/multimodality/llava.py
@@ -19,8 +19,8 @@ def main():
     max_length = max_new_tokens + prefill_length
     pretrained_model_name_or_path = "llava-hf/llava-1.5-7b-hf"
 
-    dtype = jnp.bfloat16
-    param_dtype = jnp.bfloat16
+    dtype = jnp.float16
+    param_dtype = jnp.float16
     partition_axis = ed.PartitionAxis()
 
     processor = AutoProcessor.from_pretrained(pretrained_model_name_or_path)
